add test for watch mode
obviously this is a bit crazy, hard to maintain and brittle
(particularly WRT race conditions and premature termination; cf. inline
comment) - that's why it's not a the `test` target dependency, i.e.
needs to be invoked manually

it's also testing postcss-import functionality to some extent, but it
seems okay to use that as a representative plugin